Convert all the existing layers to use the builder pattern
We have two important design desiderata:

1. We want layers in models to be configurable, so that it is possible
   to make custom architectures using composition.

This was largely archieved in Curated Transformers by letting modules
take other modules in the constructor parameters. However in Oxidized
Transformers, this solution conflicts with:

2. Layers must construct their childeren, to have control over their
   namespacing.

In Curated Transformers, we could rely on PyTorch' module abstraction
for namespacing. However, this does not work in Candle, since we have
to provide a layer's prefix when constructing it.

So, instead we settle on a design that uses builders. Rather than
directly building models, we build configuration structures (though they
could also be called uninstantiated model graphs). Each configuration
struct has a build method to create the corresponding module. This
fulfills the second desideratum by letting a module build method also
construct it's children by calling their build methods.

We fulfill the first desideratum by having each configuration take its
childern as a type-generic boxed struct. For instance, the
self-attention struct can have a generic layer norm in this way:

pub struct SelfAttentionConfig {
  layer_norm: Box<dyn BuildModule>,

  // ...
}

Any kind of layer configuration can be plugged in as the layer norm, as
long as it implements the `BuildModule` trait.

This change updates all the existing layers to this pattern. After this
change, each layer is generaly implemented as:

- A layer configuration struct.
- A `Default` implementation for the configuration.
- Setting methods following the builder pattern: allows named setting,
  but in contrast to public struct members, we can modify the struct
  fields.
- An implementation of the `BuildModule` trait, or a similar trait if
  it does not fit the signature of `BuildModule`
- A layer struct.
- Implementation of `ModuleT` or another trait.

Not all the layers have all of these parts. E.g. the rotary embeddings
layers are not fully generic. Since they are strongly bound to the
self-attention implementation, we don't expect downstream users to plug
different implementations.

The changeset also contains a small number of unrelated fixes, such as
some instances where we need to use broadcast multiplication. This is
not ideal, but since this change reimplements all the layers, it was
hard to separate these fixes.

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>