Implement `Tokenizer` and co.
The `Tokenizer` trait is the main entry point for tokenization. The
trait is implemented for `HFTokenizer` which is a wrapper around the
`tokenizers` library.

Loading is implemented through the `FromHFHub` trait, which in turn
relies on the `FromRepo` trait. The `repository` module contains a
minimal implementation of a repository system.