Better structure the KV cache, make it mutable, non-optional
Before this change, the KV cache was passed around as
`Option<Vec<KeyValueCache>>`. This has several issues. First, the use of
an option requires a lot of wrapping and unwrapping. Second, the cache
is immutable and in the future we want to pre-allocate the cache and
mutate it in-place to avoid allocations/copies.

This change removes the `Option` wrapping and makes the cache mutable.
We don't do in-place updates yet.

The type for the cache itself is now `KeyValueCache` and the cache for a
layer is `LayerKeyValueCache`. Decoders now always require passing a
`KeyValueCache` object. There is a special constructor,
`KeyValueCache::no_cache` that creates a no-op cache.

This approach allows us to treat the cache/no-cache cases in the same
way. The same operations are applied, but they are just ignored when
`no_cache` is used.

The attention mask is removed from the cache for now. We don't want to
maintain identical masks for each layer. Caching a single mask would be
an option, but it is not clear what component is responsible for
updating the mask. So we'll pass the full attention mask manually again
until we find a better design.

Co-authored-by: Madeesh Kannan <shadeMe@users.noreply.github.com>