Add `LlamaDecoder`/`LlamaCausalLM`
This change also introduces a way of checking against aggregate test vectors in
a way that is sensitive to changes in order, etc. If we take an aggregate that
is based on a commutative operator, such as addition or max, changes in the
order would still result in the same aggregates.

So rather than using a simple aggregate, we use random tensors that are
generated deterministically using a PRNG. Then we compute the dot product of the
random vector and the output as the aggregate. Since all the elements of the
random vector differ and the dot product does a pairwaise multiplication of the
output and the random vector, different orders will result in different
aggregates.