Remove `KeyValueCache::cache` hyperparameters
`KeyValueCache::cache` required specifying a number of hyperparameters,
which puts an unnecessary burden on downstream users. This change
removes the arguments. Instead we rely on the first cache update to the
key and value. We also store the layers in a `HashMap` rather than `Vec`
to remove the need to specify the number of layers ahead of time.

Eventually, we want to preallocate the cache based on the maximum
generation length, but we are not there yet. So keep it simple for now.