Fix float16 support for causal language models
Fix several dtype issues:

- When applying the logits mask, ensure that it is the right type.
- Use the minimum value for the dtype when applying the logits mask
  to avoid NaNs (as a result of -Infs).
- Convert rotary embeddings to the model's dtype.

The change also adds some tests for the float16 support. The tests are
restricted to causal language models for now, since we primarily want
to check that the models don't fail due to incorrect data types.

Furthermore, we now mask the logits before applying the test, to avoid
large variances compared to HF test vectors for masked positions.

While at it, `assert_tensor_eq` and `check_causal_lm` are converted to
macros, so that we can also use relative tolerances in the future. This is
currently not used, but I experimented with relative tolerances while
preparing this change and it would be a bit of a shame to let it whither.