Add `Nonzero` trait and implement for `Tensor`
This trait works the same as `torch.nonzero`, it returns for a given
tensor the indices of the non-zero values.

This function is currently only implemented for CPU and CUDA. I will
look into a Metal kernel later. For the time being we will only use
this for setting up flash attention, so we don't need it on Metal yet
anyway.